moved experienceBar to a new file
